Preliminary fix for group variable issue.

+++ functions.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# Copyright © 2008-2013 RAAF Technology bv
+# Copyright © 2008-2015 RAAF Technology bv
 #
 # This file is part of Session.
 #
@@ -1608,7 +1608,7 @@ function tokenReader {
             let indx+=1
         elif [ "$function" = "setVars" ]; then
             reportDebug "Writing variable \"$key\" with value: \"$value\" to $usrcfd/tmp/session.$nametmp"
-            echo "$key"=\"$value\" >> "$usrcfd/tmp/session.$nametmp"
+            echo "$key=$value" >> "$usrcfd/tmp/session.$nametmp"
         else
             reportError "Invalid function passed: $function"
             return 1
@@ -1617,7 +1617,12 @@ function tokenReader {
     if [ -e "$usrcfd/tmp/session.$nametmp" ]; then 
         reportDebug "Sourcing $usrcfd/tmp/session.$nametmp"
         source "$usrcfd/tmp/session.$nametmp"
-        rm -f "$usrcfd/tmp/session.$nametmp"
+        if [ "$debug" ]; then
+            reportDebug "Moving $usrcfd/tmp/session.$nametmp to $usrcfd/tmp/session.debug.$nametmp"
+            mv "$usrcfd/tmp/session.$nametmp" "$usrcfd/tmp/session.debug.$nametmp"
+        else
+            rm -f "$usrcfd/tmp/session.$nametmp"
+        fi
     fi
 
     typeset IFS=" "
@@ -1825,7 +1830,7 @@ function parseParameters {
         if [ "$acceptAllVariables" ] || [[ "$legalVariables" =~ " $variable " ]]; then
             value="${values[indx]}"
             reportDebug "Writing variable \"$variable\" with value: \"$value\" to $usrcfd/tmp/session.$nametmp"
-            echo "$variable"=\"$value\" >> "$usrcfd/tmp/session.$nametmp"
+            echo "$variable=$value" >> "$usrcfd/tmp/session.$nametmp"
         else
             illegalsPresent="${illegalsPresent:+$illegalsPresent }--$variable"
         fi
@@ -1834,7 +1839,12 @@ function parseParameters {
     if [ -e "$usrcfd/tmp/session.$nametmp" ]; then 
         reportDebug "Sourcing $usrcfd/tmp/session.$nametmp"
         source "$usrcfd/tmp/session.$nametmp"
-        rm -f "$usrcfd/tmp/session.$nametmp"
+        if [ "$debug" ]; then
+            reportDebug "Moving $usrcfd/tmp/session.$nametmp to $usrcfd/tmp/session.debug.$nametmp"
+            mv "$usrcfd/tmp/session.$nametmp" "$usrcfd/tmp/session.debug.$nametmp"  
+        else
+            rm -f "$usrcfd/tmp/session.$nametmp"
+        fi
     fi
 
     if [ "$illegalsPresent" ]; then
